programmer level collab time!
I figured out how to get a level made and decided why not make a programmer level collab. All you have to do is add anything that needs testing and find a way to make it to the goal. I made a loop for an example, I hope this idea goes according to plan, safe making programmers!